Tutti i problemi risolti

+++ src/main/java/it/contrader/dao/SchedaVotazioneRepository.java
@@ -7,7 +7,7 @@ import it.contrader.model.SchedaVotazione;
 
 @Repository
 @Transactional
-public interface SchedaVotazioneRepository extends CrudRepository<SchedaVotazione, Long> {
+public interface SchedaVotazioneRepository extends CrudRepository<SchedaVotazione, Integer> {
 	
 	
 }
