+++ src/main/java/it/contrader/service/ServiceDTO.java
@@ -6,12 +6,12 @@ public interface ServiceDTO<DTO> {
 
 	public List<DTO> getAll();
 
-	public DTO read(long id);
+	public DTO read(int id);
 
 	public DTO insert(DTO dto);
 
 	public DTO update(DTO dto);
 
-	public void delete(long id);
+	public void delete(int id);
 
 }
